Render merge sort in colorgrid
Treat each values in as a pixel/block within the color spectrum. If the
values are sorted, then the painted pattern will resemble what one sees
on a typical image of the color spectrum. If the values are not sorted,
then it will just look like a jumbled mess of randomly colored blocks.
The more numbers, the more pixelx/blocks can be drawn, therefore the
smoother the transition between colors will be.